Taking OpenBabel out of the equation
OpenBabel generates non-compliant MOL2 files that Dock's grid
program fails to read (Dock's dock program can read them just
fine). To fix the MOL2 files some not entirely trivial changes
have to be made. As the number of targets is small it is not
worth fixing this. Instead we require users to generate the MOL2
files and fix them by hand and store them.